refactor(tests): extract lint rules into dedicated lint.test.cjs
Moves all structural lint rules out of roadmap.test.cjs into a new
tests/lint.test.cjs, keeping domain test files focused on behaviour.
Also extracts the shared testFiles() helper to eliminate per-rule
boilerplate inside the new file. No rule logic changed.
