+++ tests/roadmap.test.cjs
@@ -1220,102 +1220,4 @@ describe('getPhaseCompletionStatus helper', () => {
   });
 });
 
-// ─────────────────────────────────────────────────────────────────────────────
-// Lint test — no bare roadmapContent.replace() in roadmap.cjs
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe('lint: no bare roadmapContent.replace() in roadmap.cjs', () => {
-  test('no bare roadmapContent.replace() outside replaceInCurrentMilestone', () => {
-    const roadmapSource = fs.readFileSync(
-      path.join(__dirname, '..', 'gsd-ng', 'bin', 'lib', 'roadmap.cjs'), 'utf-8'
-    );
-    const lines = roadmapSource.split('\n');
-    const violations = lines.filter(l =>
-      l.includes('roadmapContent.replace(') &&
-      !l.includes('replaceInCurrentMilestone') &&
-      !l.trim().startsWith('//')
-    );
-    assert.deepStrictEqual(violations, [],
-      `Bare roadmapContent.replace() found (use replaceInCurrentMilestone instead):\n${violations.join('\n')}`
-    );
-  });
-});
-
-describe('lint: no os.tmpdir() in test files (use resolveTmpDir from helpers.cjs)', () => {
-  test('no test file calls os.tmpdir() directly', () => {
-    const self = path.basename(__filename);
-    const testFiles = fs.readdirSync(__dirname).filter(f => f.endsWith('.test.cjs') && f !== self);
-    const violations = [];
-    for (const file of testFiles) {
-      const src = fs.readFileSync(path.join(__dirname, file), 'utf-8');
-      const lines = src.split('\n');
-      lines.forEach((line, i) => {
-        if (line.includes('os.tmpdir()') && !line.trim().startsWith('//')) {
-          violations.push(`${file}:${i + 1}: ${line.trim()}`);
-        }
-      });
-    }
-    assert.deepStrictEqual(violations, [],
-      `Direct os.tmpdir() usage found (use resolveTmpDir() from helpers.cjs instead):\n${violations.join('\n')}`
-    );
-  });
-});
-
-describe('lint: no inline redeclaration of helpers.cjs exports in test files', () => {
-  // Catches copy-paste of cleanup() or resolveTmpDir() instead of importing from helpers.cjs.
-  // Note: local helpers that *use* these (e.g. makeTmpDir returning {tmpDir, cleanup}) are fine.
-  // `exclude` lets us rule out legitimate forms like `const cleanup = require(...)`.
-  const forbidden = [
-    { pattern: 'function cleanup(', hint: 'import cleanup from helpers.cjs' },
-    { pattern: 'function resolveTmpDir(', hint: 'import resolveTmpDir from helpers.cjs' },
-    { pattern: 'const cleanup = ', hint: 'import cleanup from helpers.cjs', exclude: 'require(' },
-    { pattern: 'const resolveTmpDir = ', hint: 'import resolveTmpDir from helpers.cjs', exclude: 'require(' },
-  ];
-
-  for (const { pattern, hint, exclude } of forbidden) {
-    test(`no test file declares \`${pattern.trim()}\` (${hint})`, () => {
-      const self = path.basename(__filename);
-      const testFiles = fs.readdirSync(__dirname).filter(f => f.endsWith('.test.cjs') && f !== self);
-      const violations = [];
-      for (const file of testFiles) {
-        const src = fs.readFileSync(path.join(__dirname, file), 'utf-8');
-        src.split('\n').forEach((line, i) => {
-          if (
-            line.includes(pattern) &&
-            (!exclude || !line.includes(exclude)) &&
-            !line.trim().startsWith('//')
-          ) {
-            violations.push(`${file}:${i + 1}: ${line.trim()}`);
-          }
-        });
-      }
-      assert.deepStrictEqual(violations, [],
-        `Inline redeclaration of helper found:\n${violations.join('\n')}`
-      );
-    });
-  }
-});
-
-describe('lint: no bare recursive dir deletion in test files (use cleanup/cleanupSubdir from helpers.cjs)', () => {
-  test('no test file calls fs.rmSync or fs.rmdirSync with recursive:true directly', () => {
-    const self = path.basename(__filename);
-    const testFiles = fs.readdirSync(__dirname).filter(f => f.endsWith('.test.cjs') && f !== self);
-    const violations = [];
-    for (const file of testFiles) {
-      const src = fs.readFileSync(path.join(__dirname, file), 'utf-8');
-      src.split('\n').forEach((line, i) => {
-        if (
-          (line.includes('fs.rmSync(') || line.includes('fs.rmdirSync(')) &&
-          line.includes('recursive: true') &&
-          !line.trim().startsWith('//')
-        ) {
-          violations.push(`${file}:${i + 1}: ${line.trim()}`);
-        }
-      });
-    }
-    assert.deepStrictEqual(violations, [],
-      `Bare recursive dir deletion found (use cleanup() or cleanupSubdir() from helpers.cjs instead):\n${violations.join('\n')}`
-    );
-  });
-});
